requirements.txt
Cython
numpy>=1.22
scipy
scikit-image
tifffile
imageio
gdown
patool
pyunpack
scikit-image
scikit-learn
bm3d
matplotlib
tensorflow==2.13.0
cuda-python
patchify
pyyaml
plantuml
wandb>=0.15.10
streamlit
h5py